core, graph, node, runtime/wasm: Move `LinkResolver` to core
