+++ runtime/wasm/Cargo.toml
@@ -23,6 +23,7 @@ semver = "0.9.0"
 [dev-dependencies]
 graphql-parser = "0.2.0"
 parity-wasm = "0.31"
+graph-core = { path = "../../core" }
 graph-mock = { path = "../../mock" }
 # We're using the latest ipfs-api for the HTTPS support that was merged in
 # https://github.com/ferristseng/rust-ipfs-api/commit/55902e98d868dcce047863859caf596a629d10ec
